pyproject.toml
[tool.poetry]
name = "hammer-test-case"
version = "0.1.0"
description = ""
authors = ["ikhanter <khantermail@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
